requirements.txt
dotenv
sqlalchemy
pandas
apache-airflow